+++ src/Sprite.cpp
@@ -10,7 +10,7 @@ void Sprite::render() const {
   if (sprite_map)
     src = sprite_map->get_frame(animation_start_time, 0);
   else
-    src = {0.0, 0.0, (float)texture->get_width(), (float)texture->get_height()};
+    src = {0.0, 0.0, texture->get_width(), texture->get_height()};
   texture->render(&src, x, y, w, h);
 }
 
@@ -20,6 +20,6 @@ void Sprite::render(unsigned long ticks) const {
   if (sprite_map)
     src = sprite_map->get_frame(animation_start_time, ticks);
   else
-    src = {0.0, 0.0, (float)texture->get_width(), (float)texture->get_height()};
+    src = {0.0, 0.0, texture->get_width(), texture->get_height()};
   texture->render(&src, x, y, w, h);
 }
